+++ Global/GlobalConfigs.lua
@@ -46,13 +46,37 @@ function GlobalConfigs:Initialize()
             order = 1
         },
         global_alerts = {
-            name = CLM.L["DKP & Loot alerts"],
-            desc = CLM.L["Toggles alerts display when receiving DKP or loot."],
+            name = CLM.L["Point & Loot alerts"],
+            desc = CLM.L["Toggles alerts display when receiving Points or loot."],
             type = "toggle",
             set = function(i, v) self:SetAlerts(v) end,
             get = function(i) return self:GetAlerts() end,
             width = "double",
             order = 2
+        },
+        global_bindings = {
+            name = CLM.L["Bindings"],
+            desc = CLM.L["Open Key Bindings UI for AddOns"],
+            type = "execute",
+            func = (function()
+                local category = "AddOns"
+                category = _G[category] or category
+                KeyBindingFrame_LoadUI()
+                KeyBindingFrame.mode = 1
+                for _,buttons in pairs(KeyBindingFrame.categoryList.buttons) do
+                    if buttons.element then
+                        if buttons.element.name == category then
+                            HideUIPanel(InterfaceOptionsFrame)
+                            ShowUIPanel(KeyBindingFrame)
+                            KeyBindingFrame.cntCategory = buttons.element.category
+                            buttons:Click()
+                            -- OptionsList_SelectButton(buttons:GetParent(), buttons)
+                        end
+                    end
+                end
+            end),
+            width = 1,
+            order = 2.1
         },
         global_sounds = {
             name = CLM.L["Addon sounds"],
@@ -69,8 +93,17 @@ function GlobalConfigs:Initialize()
             type = "toggle",
             set = function(i, v) self:SetWoWDKPBotIntegration(v) end,
             get = function(i) return self:GetWoWDKPBotIntegration() end,
-            width = "double",
+            -- width = "double",
             order = 3
+        },
+        global_gargul_integration = {
+            name = CLM.L["Gargul Integration"],
+            desc = CLM.L["Enble Gargul integration. This will allow Gargul to take control over some aspects of CLM (starting auction from Gargul, and awarding)."],
+            type = "toggle",
+            set = function(i, v) self:SetGargulIntegration(v) end,
+            get = function(i) return self:GetGargulIntegration() end,
+            -- width = "double",
+            order = 3.5
         },
         global_wipe_ledger = {
             name = CLM.L["Wipe events"],
@@ -246,6 +279,14 @@ end
 
 function GlobalConfigs:GetWoWDKPBotIntegration()
     return self.db.wowdkpbot_integration
+end
+
+function GlobalConfigs:SetGargulIntegration(value)
+    self.db.gargul_integration = value and true or false
+end
+
+function GlobalConfigs:GetGargulIntegration()
+    return self.db.gargul_integration
 end
 
 function GlobalConfigs:SetAllowChatCommands(value)